Fixed bug in musik::core::Crypt, forgot to seed boost::random. Fixed problem with musikServer when joining Connection threads.

+++ src/cube/settings/SyncPathModel.cpp
@@ -76,7 +76,7 @@ void SyncPathModel::UpdateSyncPaths(){
     }
 
     this->SetRowCount(0);
-    this->SetRowCount(this->paths.size());
+    this->SetRowCount((int)this->paths.size());
 
 }
